Move LiteralStrGenericAlias to base event module, and make it Pydantic-compatible. Update import references
